pyproject.toml
[project]
name = "scrape-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "crawl4ai>=0.8.0",
    "ddgs>=9.10.0",
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
    "python-dotenv>=1.2.1",
]
